refactor(ray_tune): collapse files into `sweep.py` & fix w&b reports
